fix: correct DTO types for War, Aggressor, and Defender
Change isk_destroyed from int to float in Aggressor and Defender DTOs,
and finished from bool to string in War DTO to match actual ESI
API response types.

+++ src/DTO/Defender.php
@@ -11,7 +11,7 @@ class Defender implements FromArray
     public function __construct(
         public ?int $alliance_id,
         public ?int $corporation_id,
-        public int $isk_destroyed,
+        public float $isk_destroyed,
         public int $ships_killed,
     ) {}
